Make CSV generation more resilience
In production, there was test data which was not completely cleaned up. This caused a TShirtOrder (order item) to reference a parent (order id) that is null (deleted). This change improves the UX of error messages when report generation fails, makes report generation succeed in cases where values are missing, and handles any failures during report generation. Aside from this change, the problematic test data was also deleted in production